New Delete Exams and Related Booking
During client testing, it was noted that delete functionality had to be moved from the exam inventory actions dropdown list to the edit exam details modal. This functionality was also required to delete any bookings that the selected deleted exam was associated to. This functionality was tested on booked and un-booked exams that included ITA, Non-ITA and Group exams.