Update Exam Inventory Filter Color Scheme
During testing, clients recommended that the color scheme for the filter buttons on the exam inventory table be changed to the BC Government color scheme. This was achieved by applying the primary variant class to the buttons, and modifying the active button color through css class manipulation to ensure that the selected blue provided enough contrast for users. Before submitting this PR, changes were approved by clients via Trello card for this issue.