Create/Update/Delete Invigilators Per Office
Client requirements asked for new flask-admin panels for invigilators. Through these panels, users should be able to see a list of all invigilators specific to the users office (not ALL invigilators). This list of office specific invigilators should have the ability to update and delete invigilator records, and the admin panel should give the user the ability to create new invigilators as well. The last piece of required functionality would be to allow the user to search through the list of office specific invigilators based upon the invigilator name. Future work on this feature may include adding Booking information for upcoming bookings for each invigilator.