Create/Update/Delete Rooms Per Office
Client requirements asked for new flask-admin panels for rooms. Through these panels, users should be able to see a list of all rooms specific to the users office (not ALL rooms). This list of office specific rooms should have the ability to update and delete room records, and the admin panel should give the user the ability to create new rooms as well.