Create/Update/Delete Invigilators/Rooms Per Office
Client requirements asked for new flask-admin panels for invigilators and rooms. Through these panels, users should be able to see a list of all rooms or all invigilators specific to the users office (not ALL rooms or ALL invigilators). This list of office specific rooms or office specific invigilators should have the ability to update and delete room records or invigilator records, and the admin panel should give the user the ability to create new rooms and invigilators as well.