Modal/Page Bug - Press Enter Refreshes Entire Page
During client testing it was found that when any "Add Exam Modal" was
opened and the "Enter" key was pressed, the entire page would refresh.
Further testing found that this also happened on the Calendar component
Filter. To fix this issue, a event listener was added to the mounted
section of each component looking for a "Enter" key press, and to
prevent the default action of that key press. This was tested on all
reported effected components in chrome, firefox and IE.