Liaison Permission Strucutre - Back End
This PR addresses the back end changes required to impliment the new liaison permission that is required to remove the liaison role. This PR adds the migration file for the backend changes on the csr model/schema as well as the admin panel changes in order for users to add/edit this permission to other users.