Pesticide Exam Modal - Preset/Validate Exam for Pesticide Exam Type
Client testing found that the Pesticide Exam modal needed the exam type value to be preset to "Pesticide". This PR contains the fix for this issue by pre-setting the exam_type_id once the modal is opened, and validating the exam_type_id as well, such that the user can continue through the steps for creating an exam.