Need Collect Fees Yes/No
Client testing found that the Collect Fees dropdown on the edit booking modal only had the "No" value enabled, and required a "Yes" value as well. This field had to map correctly to the fees field on the booking model. This PR adds this dropdown option and also ensures that the Put method to edit a particular booking works for this field now that it has been enabled.

Client also asked that the the 3 fields (fee amount, invoice to, name of entity to bill) that were enabled when the fees option was set to yes, were also to be removed, to be possibly implemented at a later date.