Chanllenger Exams - Show if User is Office Manager
During final client testing, it was found that the challenger/monthly sessional drop down needed to be visible to users with the office manager flag set to 1. This was resolved by adding the 'is_ita_designate' getting to the logic that controls whether or not this dropdown component is visible.