Exam Types - Exam Duration with Minutes
Client testing found that pesticide exams had durations that weren't round number of hours (ie// 1h30m, 1h45m). To accomodate for this, a number_of_minutes field was added to the exam_type model/schema and then added to booking durations in the calendar, booking-modal and edit-booking-modal vue files. This was tested with the client over zoom to ensure that there was no confusion on how exams were created and handled.