Jest Tests - Booking an Exam
This PR contains the jest test responsible for booking an exam. It first creates an individual exam, then books the exam on the day after the current date. This test is meant to be run with an empty booking table, and can be run with exams in the exam inventory table. To run this test type `npm test -t bookings-book-exam` in command line.