Phase 1 - Notification Banner Quick Filter Button
Client required that the notification banner for group/individual exams had the ability to re-direct the user to the exam-inventory page WITH preset inventory filters set. This PR has a button added to the rightmost corner of the notification banner, that when clicked, pushes the user to the exam route, and sets the exam filter mutation to specific values based upon which exam notification is present (group, individual or both).