Phase 2 - INVIGILATORS! Many-to-many for Group Exams and Shadow Invigilators for all exams
MANY INVIGILATORS:

Client required the ability to add many invigilators to an exam if there are more than X students per exam. This feature required moving the invigilator FK off the booking model and creating a secondary many-to-many relationship table for booking invigilators.

Testing:
	- all individual exams tested by creating, booking, scheduling invigilators and sbc_staff
	- all group exams tested by creating, booking, scheduling
	- all bookings booked and correct times and displayed correct invigilators
	- all bookings responded correctly to having their invigilator changed
	- all bookings responded correctly to having invigilators decreased
	- all exams displayed correct information for invigilators on the office agenda (one or many)

Client Feedback Changes:

Bugs:

SHADOW INVIGILATORS:

Client required a net new feature to track invigilator training (ie// shadow invigilators). The mapping of shadow invigilator to booking will be 1:1. In order to accomplish this task, front end changes were made to the edit booking and edit group exam modals such that a user can add shadow invigilators to both individual and group exams, as well as change or remove the shadow invigilator once one has been set. Other requirements for this feature were to make changes to the invigilator admin panel such that number of times an invigilator had been added as a shadow invigilator to a booking was being tracked, and once they've completed their training to set a flag telling a given user that this was the case. Throughout the booking process now, invigilators will only show up in invigilator lists IFF they have completed 2 shadow sessions, otherwise they will appear in the shadow invigilator lists until this has been completed.

The final requirement for this feature was to add shadow invigilators to the exam export list csv function, such that if there was a shadow invigilator for a given exam, that their name will be listed.

This feature requires the back-end to be rebuilt using the normal process for flask. Changes should show that there is a shadow_invigilator_id field on the booking model, and a shadow_count and shadow_flag field on the invigilator model. Appropriate changes were made to the schemas to reflect these changes.

Client Feedback Changes:
	- clicking the "Remove Invigilator" Yes button now triggers a full modal submit, instead of having to click "Yes" then "Submit"
Testing:
	- all individual exams tested by creating, booking and selecting one shadow invigilator
	- all group exams tested by creating, booking and selecting one shadow invigilator
	- all information correctly displaying on the office agenda
Bugs:
	- client noted that adding and removing shadow invigilators wasn't correctly adding/subtracting from the invigilator shadow_count field - this has been resolved. adding/removing shadow invigilators from individual/group exams has been tested and is now working correctly