Phase 2 - Shadow Invigilators
Client required a net new feature to track invigilator training (ie// shadow invigilators). The mapping of shadow invigilator to booking will be 1:1. In order to accomplish this task, front end changes were made to the edit booking and edit group exam modals such that a user can add shadow invigilators to both individual and group exams, as well as change or remove the shadow invigilator once one has been set. Other requirements for this feature were to make changes to the invigilator admin panel such that number of times an invigilator had been added as a shadow invigilator to a booking was being tracked, and once they've completed their training to set a flag telling a given user that this was the case. Throughout the booking process now, invigilators will only show up in invigilator lists IFF they have completed 2 shadow sessions, otherwise they will appear in the shadow invigilator lists until this has been completed.

The final requirement for this feature was to add shadow invigilators to the exam export list csv function, such that if there was a shadow invigilator for a given exam, that their name will be listed.

This feature requires the back-end to be rebuilt using the normal process for flask. Changes should show that there is a shadow_invigilator_id field on the booking model, and a shadow_count and shadow_flag field on the invigilator model. Appropriate changes were made to the schemas to reflect these changes.