Recurring Events Edit/Delete Series - Appointments
Client requires functionality in order to edit and delete a series of recurring events. This commit edits functionality behind the appointments calendar/modals in order to do this.

Two new end points were created in order to accomplish. Both require a uuid to be provided, which is now generated on creation and submitted on creation to the updated appointment model.