Recurring Events Edit/Delete Series - Booking
Client required functionality to edit and delete recurring events on the booking calendar(non-exam and blackouts). This required adding new end points to the booking package for editing and deleting recurring events.

Front end development shows a new row on the edit booking modal, only visible if the booking event has a recurring_uuid associated with it. This row houses a button to delete the recurring series, and a toggle button for editing
the series of events or a single event.